+++ backend/src/main/java/com/pickeat/backend/restaurant/application/WishRestaurantSearchService.java
@@ -6,8 +6,8 @@ import com.pickeat.backend.restaurant.application.dto.request.RestaurantRequest;
 import com.pickeat.backend.restaurant.application.dto.request.WishRestaurantRequest;
 import com.pickeat.backend.room.domain.Room;
 import com.pickeat.backend.room.domain.repository.RoomRepository;
-import com.pickeat.backend.tobe.wish.domain.Wish;
-import com.pickeat.backend.tobe.wish.domain.repository.WishRepository;
+import com.pickeat.backend.wish.domain.Wish;
+import com.pickeat.backend.wish.domain.repository.WishRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
