--- backend/src/main/java/com/pickeat/backend/tobe/wish/application/dto/request/ImageRequest.java
@@ -1,5 +1,0 @@
-package com.pickeat.backend.tobe.wish.application.dto.request;
-
-public record ImageRequest(String key, String downloadUrl) {
-
-}
