+++ backend/src/main/java/com/pickeat/backend/wish/domain/Wish.java
@@ -1,88 +1,39 @@
 package com.pickeat.backend.wish.domain;
 
 import com.pickeat.backend.global.BaseEntity;
-import com.pickeat.backend.restaurant.domain.FoodCategory;
-import jakarta.persistence.CascadeType;
+import com.pickeat.backend.restaurant.domain.RestaurantInfo;
+import com.pickeat.backend.room.domain.Room;
 import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
 @Getter
+@Entity(name = "wish_v2")
+@Table(name = "wish_v2")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Wish extends BaseEntity {
 
-    @Column(nullable = false)
-    private String name;
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "room_id", nullable = false)
+    private Room room;
 
     @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private FoodCategory foodCategory;
+    @Embedded
+    private RestaurantInfo restaurantInfo;
 
-    @Column(nullable = false)
-    private String roadAddressName;
-
-    private String tags;
-
-    private String placeUrl;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "wish_list_id", nullable = false)
-    private WishList wishList;
-
-    @OneToMany(mappedBy = "wish", cascade = CascadeType.REMOVE)
-    private final List<WishPicture> wishPictures = new ArrayList<>();
-
-    public Wish(
-            String name,
-            FoodCategory foodCategory,
-            String roadAddressName,
-            String tags,
-            String placeUrl,
-            WishList wishList
-    ) {
-        this.name = name;
-        this.foodCategory = foodCategory;
-        this.roadAddressName = roadAddressName;
-        this.tags = convertEmptyTagsToNull(tags);
-        this.placeUrl = placeUrl;
-        this.wishList = wishList;
+    public Wish(Room room, RestaurantInfo restaurantInfo) {
+        this.room = room;
+        this.restaurantInfo = restaurantInfo;
     }
 
-    public void updateName(String name) {
-        this.name = name;
-    }
-
-    public void updateFoodCategory(FoodCategory foodCategory) {
-        this.foodCategory = foodCategory;
-    }
-
-    public void updateRoadAddressName(String roadAddressName) {
-        this.roadAddressName = roadAddressName;
-    }
-
-    public void updateTags(String tags) {
-        this.tags = convertEmptyTagsToNull(tags);
-    }
-
-    public void updatePlaceUrl(String placeUrl) {
-        this.placeUrl = placeUrl;
-    }
-
-    private String convertEmptyTagsToNull(String tags) {
-        if (tags == null || tags.isBlank()) {
-            return null;
-        }
-        return tags;
+    public void updateRestaurantInfo(RestaurantInfo restaurantInfo) {
+        this.restaurantInfo = restaurantInfo;
     }
 }
