+++ backend/src/main/java/com/pickeat/backend/wish/ui/WishController.java
@@ -4,7 +4,6 @@ import com.pickeat.backend.global.auth.annotation.LoginUserId;
 import com.pickeat.backend.global.log.BusinessLogging;
 import com.pickeat.backend.wish.application.WishService;
 import com.pickeat.backend.wish.application.dto.request.WishRequest;
-import com.pickeat.backend.wish.application.dto.request.WishUpdateRequest;
 import com.pickeat.backend.wish.application.dto.response.WishResponse;
 import com.pickeat.backend.wish.ui.api.WishApiSpec;
 import jakarta.validation.Valid;
@@ -16,14 +15,12 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@Deprecated
-@RestController
-@RequestMapping("/api/v1")
+@RestController("WishControllerV2")
+@RequestMapping("/api/v2")
 @RequiredArgsConstructor
 public class WishController implements WishApiSpec {
 
@@ -31,13 +28,13 @@ public class WishController implements WishApiSpec {
 
     @Override
     @BusinessLogging("위시 생성")
-    @PostMapping(value = "/wishLists/{wishListId}/wishes")
+    @PostMapping(value = "/rooms/{roomId}/wishes")
     public ResponseEntity<WishResponse> createWish(
-            @PathVariable("wishListId") Long wishListId,
+            @PathVariable("roomId") Long roomId,
             @Valid @RequestBody WishRequest request,
             @LoginUserId Long userId
     ) {
-        WishResponse wishResponse = wishService.createWish(wishListId, request, userId);
+        WishResponse wishResponse = wishService.createWish(roomId, request, userId);
         return ResponseEntity
                 .status(HttpStatus.CREATED)
                 .body(wishResponse);
@@ -52,33 +49,12 @@ public class WishController implements WishApiSpec {
     }
 
     @Override
-    @BusinessLogging("위시 수정")
-    @PutMapping("/wishes/{wishId}")
-    public ResponseEntity<WishResponse> updateWish(
-            @PathVariable("wishId") Long wishId,
-            @Valid @RequestBody WishUpdateRequest request,
-            @LoginUserId Long userId
-    ) {
-        WishResponse updated = wishService.updateWish(wishId, userId, request);
-        return ResponseEntity.ok(updated);
-    }
-
-    @Override
-    @GetMapping("/wishLists/{wishListId}/wishes")
-    public ResponseEntity<List<WishResponse>> getWishesInWishList(
-            @PathVariable("wishListId") Long wishListId,
+    @GetMapping("/rooms/{roomId}/wishes")
+    public ResponseEntity<List<WishResponse>> getWishesInRoom(
+            @PathVariable("roomId") Long roomId,
             @LoginUserId Long loginUserId
     ) {
-        List<WishResponse> wishes = wishService.getWishes(wishListId, loginUserId);
-        return ResponseEntity.ok(wishes);
-    }
-
-    @Override
-    @GetMapping("/wishLists/templates/{wishListId}/wishes")
-    public ResponseEntity<List<WishResponse>> getWishesInTemplates(
-            @PathVariable("wishListId") Long wishListId
-    ) {
-        List<WishResponse> wishes = wishService.getWishesFromTemplates(wishListId);
+        List<WishResponse> wishes = wishService.getWishes(roomId, loginUserId);
         return ResponseEntity.ok(wishes);
     }
 }
