+++ backend/src/main/java/com/pickeat/backend/wish/ui/api/WishPictureApiSpec.java
@@ -5,14 +5,18 @@ import com.pickeat.backend.wish.application.dto.response.WishPictureResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.constraints.NotEmpty;
-import java.util.List;
+import jakarta.validation.constraints.NotNull;
+import org.springframework.http.MediaType;
+import org.springframework.http.ProblemDetail;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -38,19 +42,66 @@ public interface WishPictureApiSpec {
                     description = "잘못된 요청 데이터 (부적절한 이미지 타입)",
                     content = @Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = org.springframework.http.ProblemDetail.class),
-                            examples = @io.swagger.v3.oas.annotations.media.ExampleObject(
-                                    name = "부적절한 이미지 타입",
-                                    value = "{\"type\":\"about:blank\",\"title\":\"Bad Request\",\"status\":400,\"detail\":\"이미지 파일만 업로드할 수 있습니다.\",\"instance\":\"/api/v1/wish/1/wishpictures\"}"
+                            schema = @Schema(implementation = ProblemDetail.class),
+                            examples = @ExampleObject(
+                                    name = "INVALID_IMAGE_TYPE",
+                                    value = """
+                                            {
+                                              "type": "about:blank",
+                                              "title": "INVALID_IMAGE_TYPE",
+                                              "status": 400,
+                                              "detail": "이미지 파일만 업로드할 수 있습니다.",
+                                              "instance": "/api/v1/wish/1/wishpictures"
+                                            }
+                                            """
+                            )
+                    )
+            ),
+            @ApiResponse(
+                    responseCode = "401",
+                    description = "인증되지 않은 사용자",
+                    content = @Content(
+                            mediaType = "application/json",
+                            schema = @Schema(implementation = ProblemDetail.class),
+                            examples = @ExampleObject(
+                                    name = "UNAUTHORIZED",
+                                    value = """
+                                            {
+                                                "type": "about:blank",
+                                                "title": "UNAUTHORIZED",
+                                                "status": 401,
+                                                "detail": "인증 정보가 유효하지 않습니다.",
+                                                "instance": "/api/v1/wish/1/wishpictures"
+                                            }
+                                            """
+                            )
+                    )
+            ),
+            @ApiResponse(
+                    responseCode = "404",
+                    description = "위시를 찾을 수 없음",
+                    content = @Content(
+                            mediaType = "application/json",
+                            schema = @Schema(implementation = ProblemDetail.class),
+                            examples = @ExampleObject(
+                                    name = "WISH_NOT_FOUND",
+                                    value = """
+                                            {
+                                                "type": "about:blank",
+                                                "title": "WISH_NOT_FOUND",
+                                                "status": 404,
+                                                "detail": "위시를 찾을 수 없습니다.",
+                                                "instance": "/api/v1/wish/1/wishpictures"
+                                            }
+                                            """
                             )
                     )
             )
     })
-    ResponseEntity<List<WishPictureResponse>> createWishPictures(
-            @Parameter(description = "위시 ID", example = "1")
-            @PathVariable("wishId") Long wishId,
-            @Parameter(description = "위시 사진 목록")
-            @RequestPart("wishPictures") @NotEmpty List<MultipartFile> wishPictures,
+    @PostMapping(value = "/wish/{wishId}/wishpictures", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    ResponseEntity<WishPictureResponse> createWishPictures(
+            @Parameter(description = "위시 ID", example = "1") @PathVariable("wishId") Long wishId,
+            @Parameter(description = "위시 사진 목록") @RequestPart("wishPictures") @NotNull MultipartFile wishPictures,
             @Parameter(hidden = true) @LoginUserId Long userId
     );
 
@@ -62,29 +113,50 @@ public interface WishPictureApiSpec {
     @ApiResponses(value = {
             @ApiResponse(responseCode = "204", description = "위시 사진 삭제 성공"),
             @ApiResponse(
+                    responseCode = "401",
+                    description = "인증되지 않은 사용자",
+                    content = @Content(
+                            mediaType = "application/json",
+                            schema = @Schema(implementation = ProblemDetail.class),
+                            examples = @ExampleObject(
+                                    name = "UNAUTHORIZED",
+                                    value = """
+                                            {
+                                                "type": "about:blank",
+                                                "title": "UNAUTHORIZED",
+                                                "status": 401,
+                                                "detail": "인증 정보가 유효하지 않습니다.",
+                                                "instance": "/api/v1/wish/1/wishpictures"
+                                            }
+                                            """
+                            )
+                    )
+            ),
+            @ApiResponse(
                     responseCode = "404",
                     description = "위시를 찾을 수 없음",
                     content = @Content(
                             mediaType = "application/json",
-                            schema = @Schema(implementation = org.springframework.http.ProblemDetail.class),
-                            examples = @io.swagger.v3.oas.annotations.media.ExampleObject(
-                                    name = "위시를 찾을 수 없음",
+                            schema = @Schema(implementation = ProblemDetail.class),
+                            examples = @ExampleObject(
+                                    name = "WISH_NOT_FOUND",
                                     value = """
                                             {
-                                                "type":"about:blank",
-                                                "title":"Not Found",
-                                                "status":404,
-                                                "detail":"위시를 찾을 수 없습니다.",
-                                                "instance":"/api/v1/wish/1/wishpictures"
+                                                "type": "about:blank",
+                                                "title": "WISH_NOT_FOUND",
+                                                "status": 404,
+                                                "detail": "위시를 찾을 수 없습니다.",
+                                                "instance": "/api/v1/wish/1/wishpictures"
                                             }
                                             """
                             )
                     )
             )
     })
+    @DeleteMapping("/wish/{wishId}/wishpictures")
     ResponseEntity<Void> deleteWishPictures(
-            @Parameter(description = "위시 ID", example = "1")
-            @PathVariable("wishId") Long wishId,
+            @Parameter(description = "위시 ID", example = "1") @PathVariable("wishId") Long wishId,
             @Parameter(hidden = true) @LoginUserId Long userId
     );
 }
+
