+++ backend/src/test/java/com/pickeat/backend/tobe/acceptance_test/scenario/PickeatByWishScenarioTest.java
@@ -42,9 +42,9 @@ import com.pickeat.backend.restaurant.domain.FoodCategory;
 import com.pickeat.backend.room.application.dto.request.RoomInvitationRequest;
 import com.pickeat.backend.room.application.dto.request.RoomRequest;
 import com.pickeat.backend.room.application.dto.response.RoomResponse;
-import com.pickeat.backend.tobe.wish.application.dto.request.WishRequest;
-import com.pickeat.backend.tobe.wish.application.dto.response.WishResponse;
 import com.pickeat.backend.user.application.dto.UserResponse;
+import com.pickeat.backend.wish.application.dto.request.WishRequest;
+import com.pickeat.backend.wish.application.dto.response.WishResponse;
 import io.restassured.RestAssured;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
