+++ backend/src/test/java/com/pickeat/backend/tobe/wish/application/WishPictureServiceTest.java
@@ -17,12 +17,14 @@ import com.pickeat.backend.room.domain.RoomUser;
 import com.pickeat.backend.room.domain.repository.RoomRepository;
 import com.pickeat.backend.room.domain.repository.RoomUserRepository;
 import com.pickeat.backend.tobe.fixture.WishFixture;
-import com.pickeat.backend.tobe.wish.application.dto.response.WishPictureResponse;
-import com.pickeat.backend.tobe.wish.domain.Wish;
-import com.pickeat.backend.tobe.wish.domain.repository.WishRepository;
-import com.pickeat.backend.tobe.wish.infrastructure.LocalImageUploadClient;
 import com.pickeat.backend.user.domain.User;
 import com.pickeat.backend.user.domain.repository.UserRepository;
+import com.pickeat.backend.wish.application.ImageUploadClient;
+import com.pickeat.backend.wish.application.WishPictureService;
+import com.pickeat.backend.wish.application.dto.response.WishPictureResponse;
+import com.pickeat.backend.wish.domain.Wish;
+import com.pickeat.backend.wish.domain.repository.WishRepository;
+import com.pickeat.backend.wish.infrastructure.LocalImageUploadClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
