+++ backend/src/test/java/com/pickeat/backend/tobe/wish/application/WishServiceTest.java
@@ -13,12 +13,13 @@ import com.pickeat.backend.restaurant.domain.RestaurantInfo;
 import com.pickeat.backend.room.domain.Room;
 import com.pickeat.backend.room.domain.RoomUser;
 import com.pickeat.backend.tobe.fixture.WishFixture;
-import com.pickeat.backend.tobe.wish.application.dto.request.WishRequest;
-import com.pickeat.backend.tobe.wish.application.dto.request.WishUpdateRequest;
-import com.pickeat.backend.tobe.wish.application.dto.response.WishResponse;
-import com.pickeat.backend.tobe.wish.domain.Wish;
-import com.pickeat.backend.tobe.wish.domain.repository.WishRepository;
 import com.pickeat.backend.user.domain.User;
+import com.pickeat.backend.wish.application.WishService;
+import com.pickeat.backend.wish.application.dto.request.WishRequest;
+import com.pickeat.backend.wish.application.dto.request.WishUpdateRequest;
+import com.pickeat.backend.wish.application.dto.response.WishResponse;
+import com.pickeat.backend.wish.domain.Wish;
+import com.pickeat.backend.wish.domain.repository.WishRepository;
 import java.util.Comparator;
 import java.util.List;
 import org.junit.jupiter.api.Nested;
